+++ src/service/transformer/declaration-bundler/util/merge-exports/merge-exports.ts
@@ -9,8 +9,9 @@ const EMPTY_MODULE_SPECIFIER_TOKEN = "_#gen__empty__module__specifier";
 export function mergeExports(statements: TS.Statement[], typescript: typeof TS): TS.Statement[] {
 	const exports = statements.filter(typescript.isExportDeclaration);
 	const otherStatements = statements.filter(statement => !typescript.isExportDeclaration(statement));
-	const moduleSpecifierToExportedBindingsMap: Map<string, Set<string>> = new Map();
+	const moduleSpecifierToAliasedExportedBindings: Map<string, Map<string, Set<string>>> = new Map();
 	const exportDeclarations: Set<TS.ExportDeclaration> = new Set();
+	const reExportedSpecifiers = new Set<string>();
 
 	for (const exportDeclaration of exports) {
 		// If the ModuleSpecifier is given and it isn't a string literal, leave it as it is
@@ -21,48 +22,60 @@ export function mergeExports(statements: TS.Statement[], typescript: typeof TS):
 
 		const specifierText = exportDeclaration.moduleSpecifier == null ? EMPTY_MODULE_SPECIFIER_TOKEN : exportDeclaration.moduleSpecifier.text;
 
-		let exportedBindings = moduleSpecifierToExportedBindingsMap.get(specifierText);
-		if (exportedBindings == null) {
-			exportedBindings = new Set();
-			moduleSpecifierToExportedBindingsMap.set(specifierText, exportedBindings);
+		let aliasedExportedBindings = moduleSpecifierToAliasedExportedBindings.get(specifierText);
+
+		if (aliasedExportedBindings == null) {
+			aliasedExportedBindings = new Map();
+			moduleSpecifierToAliasedExportedBindings.set(specifierText, aliasedExportedBindings);
 		}
 
 		if (exportDeclaration.exportClause != null) {
-			const aliasedExportSpecifiers: Set<TS.ExportSpecifier> = new Set();
-
 			// Take all aliased exports
 			for (const element of exportDeclaration.exportClause.elements) {
-				if (element.propertyName != null && element.name != null) {
-					aliasedExportSpecifiers.add(element);
-				} else {
-					exportedBindings.add(element.name.text);
+				const propertyName = element.propertyName != null ? element.propertyName.text : element.name.text;
+				const alias = element.name.text;
+				let setForExportedBinding = aliasedExportedBindings.get(propertyName);
+				if (setForExportedBinding == null) {
+					setForExportedBinding = new Set();
+					aliasedExportedBindings.set(propertyName, setForExportedBinding);
 				}
+				setForExportedBinding.add(alias);
 			}
-
-			// If at least 1 is aliased, generate an export containing only those
-			if (aliasedExportSpecifiers.size > 0) {
-				exportDeclarations.add(
-					typescript.createExportDeclaration(
-						undefined,
-						undefined,
-						typescript.createNamedExports([...aliasedExportSpecifiers]),
-						exportDeclaration.moduleSpecifier
-					)
-				);
-			}
-		} else {
+		}
+		// If it has no exportClause, it's a reexport (such as export * from "./<specifier>").
+		else {
+			// Don't include the same clause twice
+			if (reExportedSpecifiers.has(specifierText)) continue;
+			reExportedSpecifiers.add(specifierText);
 			exportDeclarations.add(exportDeclaration);
 		}
 	}
 
-	for (const [specifier, exportedBindings] of moduleSpecifierToExportedBindingsMap) {
+	for (const [specifier, exportedBindings] of moduleSpecifierToAliasedExportedBindings) {
 		if (exportedBindings.size === 0) continue;
+
+		const exportSpecifiers: TS.ExportSpecifier[] = [];
+		const bindings = new Set<string>();
+
+		for (const [propertyName, aliases] of exportedBindings) {
+			for (const alias of aliases) {
+				// If a binding, A, is exported already, it cannot be exported again.
+				if (bindings.has(alias)) continue;
+				bindings.add(alias);
+
+				if (propertyName === alias) {
+					exportSpecifiers.push(typescript.createExportSpecifier(undefined, alias));
+				} else {
+					exportSpecifiers.push(typescript.createExportSpecifier(propertyName, alias));
+				}
+			}
+		}
 
 		exportDeclarations.add(
 			typescript.createExportDeclaration(
 				undefined,
 				undefined,
-				typescript.createNamedExports([...exportedBindings].map(exportedBinding => typescript.createExportSpecifier(undefined, exportedBinding))),
+				typescript.createNamedExports(exportSpecifiers),
 				specifier === EMPTY_MODULE_SPECIFIER_TOKEN ? undefined : typescript.createStringLiteral(ensureHasLeadingDotAndPosix(specifier))
 			)
 		);