+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.1.83](https://github.com/wessberg/rollup-plugin-ts/compare/v1.1.82...v1.1.83) (2019-11-28)
+
+### Bug Fixes
+
+- **bug:** fix multiple aliased exports from same module ([db29b86](https://github.com/wessberg/rollup-plugin-ts/commit/db29b86c0714c81a0053b92ac071e9cba6c86893))
+
+## [1.1.82](https://github.com/wessberg/rollup-plugin-ts/compare/v1.1.81...v1.1.82) (2019-11-28)
+
+### Bug Fixes
+
+- **bug:** fix multiple aliased exports from same module ([d7e7ec5](https://github.com/wessberg/rollup-plugin-ts/commit/d7e7ec578c8e0f56ab8b9aeaff4a2db3c5d4f5d5))
+
+## [1.1.81](https://github.com/wessberg/rollup-plugin-ts/compare/v1.1.80...v1.1.81) (2019-11-28)
+
+### Bug Fixes
+
+- **bug:** fix with multiple duplicate reexports ([294d7c5](https://github.com/wessberg/rollup-plugin-ts/commit/294d7c519e723260eb9378fd6d0b307947f95446))
+
 ## [1.1.80](https://github.com/wessberg/rollup-plugin-ts/compare/v1.1.79...v1.1.80) (2019-11-17)
 
 ### Bug Fixes